+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Release Notes for Blitz
 
 ## 3.6.12 - Unreleased
+### Added
+- Added the API token authentication method to the Cloudflare purger ([#263](https://github.com/putyourlightson/craft-blitz/issues/263)).
+
 ### Fixed
 - Fixed an error that could occur if the CP trigger was set to `null` ([#264](https://github.com/putyourlightson/craft-blitz/issues/264)).